+++ libs/dimcli/cli.h
@@ -334,15 +334,26 @@ public:
     void printOptions(std::ostream & os, const std::string & cmd = {});
     void printCommands(std::ostream & os);
 
+    // If !exitCode() prints the errMsg and errDetail (if present), but does
+    // nothing if exitCode() is EX_OK. Returns exitCode(). Only makes sense 
+    // after parsing has completed.
+    int printError(std::ostream & oerr);
+
     //-----------------------------------------------------------------------
     // Parsing
 
+    // Parse the command line, populate the options, and set the error and 
+    // other miscellaneous state.
+    //
+    // The ostream& argument is only used to print the error message, if any,
+    // via printError(). Error information can also be extracted after parse()
+    // completes, see errMsg() and friends.
     bool parse(size_t argc, char * argv[]);
-    bool parse(std::ostream & os, size_t argc, char * argv[]);
+    bool parse(std::ostream & oerr, size_t argc, char * argv[]);
 
     // "args" is non-const so response files can be expanded in place.
     bool parse(std::vector<std::string> & args);
-    bool parse(std::ostream & os, std::vector<std::string> & args);
+    bool parse(std::ostream & oerr, std::vector<std::string> & args);
 
     // Sets all options to their defaults, called internally when parsing
     // starts.
@@ -436,6 +447,13 @@ public:
     // action of the empty "" command, which can be set via cli.action() just 
     // like any other command.
     bool exec();
+    bool exec(std::ostream & oerr);
+
+    // Helpers to parse and, if successful, execute.
+    bool exec(size_t argc, char * argv[]);
+    bool exec(std::ostream & oerr, size_t argc, char * argv[]);
+    bool exec(std::vector<std::string> & args);
+    bool exec(std::ostream & oerr, std::vector<std::string> & args);
 
     // Sets exitCode(), errMsg(), and errDetail(), intended to be called from
     // command actions, parsing related failures should use badUsage().